+++ core/app_shared.go
@@ -19,8 +19,8 @@ import (
 	"core-building-block/driven/storage"
 	"core-building-block/utils"
 
-	"github.com/rokwire/logging-library-go/v2/errors"
-	"github.com/rokwire/logging-library-go/v2/logutils"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/errors"
+	"github.com/rokwire/rokwire-building-block-sdk-go/utils/logging/logutils"
 )
 
 func (app *application) sharedGetAppConfig(appTypeIdentifier string, orgID *string, versionNumbers model.VersionNumbers, apiKey *string, admin bool) (*model.ApplicationConfig, error) {